fix: promote to error
