+++ crates/diagnostics/src/lint.rs
@@ -270,17 +270,6 @@ pub fn unsigned_comparison(span: &Span, always_true: bool) -> LisetteDiagnostic 
         )
 }
 
-pub fn nan_comparison(span: &Span, always_true: bool) -> LisetteDiagnostic {
-    let result = if always_true { "true" } else { "false" };
-
-    LisetteDiagnostic::warn(format!("Comparison with NaN is always {result}"))
-        .with_lint_code("nan_comparison")
-        .with_span_label(span, format!("always {result}"))
-        .with_help(
-            "Per IEEE 754, NaN compares unequal to every value including itself. Use `math.IsNaN(x)` to test for NaN.",
-        )
-}
-
 pub fn verbose_failure_propagation(span: &Span) -> LisetteDiagnostic {
     LisetteDiagnostic::warn("Verbose failure propagation")
         .with_lint_code("verbose_failure_propagation")
