+++ crates/semantics/src/passes/lints/ast_walk/checks/mod.rs
@@ -8,7 +8,6 @@ mod identical_if_branches;
 mod invisible_in_string;
 mod match_literal_collection;
 mod naming;
-mod nan_comparison;
 mod replaceable_with_zero_fill;
 mod rest_only_slice_pattern;
 mod self_assignment;
@@ -30,7 +29,6 @@ pub use invisible_in_string::{
 };
 pub use match_literal_collection::check_match_literal_collection;
 pub use naming::{check_expression_naming, check_pattern_naming};
-pub use nan_comparison::check_nan_comparison;
 pub use replaceable_with_zero_fill::check_replaceable_with_zero_fill;
 pub use rest_only_slice_pattern::check_rest_only_slice_pattern;
 pub use self_assignment::check_self_assignment;
